+++ src/main/java/io/woori/account/wooriaccount/role/repository/jpa/RoleRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.Repository;
 import io.woori.account.wooriaccount.role.domain.Role;
 import io.woori.account.wooriaccount.role.repository.basic.CommonRoleRepository;
 
-public interface RoleRepository extends Repository<Role, Long>, CommonRoleRepository {
+public interface RoleRepository extends Repository<Role, Byte>, CommonRoleRepository {
 
 }
